refactor(guestbook): one source of truth for the field limits
src/lib/constants.ts declared MAX_NAME_LENGTH = 40, MAX_MESSAGE_LENGTH = 280
and its own GuestbookEntry, none of which anything imported — while the live
values (32 / 500) were written out three more times, in lib/guestbook.ts and
in both guestbook islands. Its header comment claimed the islands imported
from it, which they didn't. Anyone who trusted the constants would have got a
form that caps messages at 280 in front of a server that truncates at 500.

constants.ts now holds the real values and the entry type; guestbook.ts
re-exports them so existing imports keep working, and both islands import
them instead of re-declaring. No behaviour change — both skins still render
maxLength 32 / 500 / 200, verified in the built output.
